JNG-4321 Keep the order of joins is there is no explicit dependency
