+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/model/RdbmsCount.java
@@ -20,7 +20,6 @@ package hu.blackbelt.judo.runtime.core.dao.rdbms.query.model;
  * #L%
  */
 
-import hu.blackbelt.judo.dao.api.DAO;
 import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.*;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.StatementExecutor;
@@ -31,7 +30,6 @@ import hu.blackbelt.mapper.api.Coercer;
 import lombok.Builder;
 import lombok.NonNull;
 import org.eclipse.emf.common.util.*;
-import org.eclipse.emf.ecore.EAttribute;
 import org.eclipse.emf.ecore.EClass;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 
@@ -202,11 +200,11 @@ public class RdbmsCount<ID> {
     }
 
     private String getJoin(String prefix, Coercer coercer, MapSqlParameterSource sqlParameters, EMap<Node, String> newPrefixes, RdbmsJoin firstJoin) {
-        String join = joins.stream().collect(Collectors.toMap(j -> j, j -> j.toSql(prefix, coercer, sqlParameters, newPrefixes, from == null && Objects.equals(j, firstJoin)))).entrySet()
-                              .stream().sorted((l, r) -> RdbmsJoin.RDBMS_JOIN_COMPARATOR.compare(l.getKey(), r.getKey()))
-                              .map(Map.Entry::getValue)
-                              .collect(Collectors.joining());
-        return join;
+        Map<RdbmsJoin, String> joinMap = joins.stream().collect(Collectors.toMap(j -> j, j -> j.toSql(prefix, coercer, sqlParameters, newPrefixes, from == null && Objects.equals(j, firstJoin))));
+        return joins.stream()
+                    .sorted(new RdbmsJoinComparator(joins))
+                    .map(joinMap::get)
+                    .collect(Collectors.joining());
     }
 
 }
