+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/model/join/RdbmsJoinComparator.java
@@ -1,0 +1,54 @@
+package hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join;
+
+import java.util.*;
+
+public class RdbmsJoinComparator implements Comparator<RdbmsJoin> {
+
+    final List<RdbmsJoin> originalOrder;
+
+    public RdbmsJoinComparator(Collection<RdbmsJoin> originalOrder) {
+        this.originalOrder = new ArrayList<>(originalOrder);
+    }
+
+    @Override
+    public int compare(RdbmsJoin left, RdbmsJoin right) {
+        checkArguments(left, right);
+
+        String leftAliasToCompare = left.alias;
+        if (left.aliasToCompareWith != null && !left.aliasToCompareWith.isBlank()) {
+            leftAliasToCompare = left.aliasToCompareWith;
+        }
+
+        String rightAliasToCompare = left.alias;
+        if (right.aliasToCompareWith != null && !right.aliasToCompareWith.isBlank()) {
+            rightAliasToCompare = right.aliasToCompareWith;
+        }
+
+        if (left.equals(right)) {
+            return 0;
+        }
+
+        // left depends on right
+        if (right.joinConditionPartnerTableAliases.contains(leftAliasToCompare)) {
+            return -1;
+        }
+
+        // right depends on left
+        if (left.joinConditionPartnerTableAliases.contains(rightAliasToCompare)) {
+            return 1;
+        }
+
+        // there is no explicit dependency but original order must be kept
+        return originalOrder.indexOf(left) - originalOrder.indexOf(right);
+    }
+
+    private static void checkArguments(RdbmsJoin left, RdbmsJoin right) {
+        if (left == null || right == null) {
+            throw new IllegalArgumentException("Compared arguments cannot be null");
+        }
+        if (left.alias == null || left.alias.isBlank() || right.alias == null || right.alias.isBlank()) {
+            throw new IllegalArgumentException("Compared arguments' aliases cannot be null/empty");
+        }
+    }
+
+}
