+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/model/join/RdbmsNavigationJoin.java
@@ -272,14 +272,14 @@ public class RdbmsNavigationJoin<ID> extends RdbmsJoin {
     }
 
     private String getJoin(Coercer coercer, MapSqlParameterSource sqlParameters, String subSelectJoinPrefix, EMap<Node, String> newPrefixes) {
-        String join = subJoins.stream().collect(Collectors.toMap(j -> j, j -> j.toSql(subSelectJoinPrefix, coercer, sqlParameters, newPrefixes, false))).entrySet().stream()
-                                 .sorted((l, r) -> RdbmsJoin.RDBMS_JOIN_COMPARATOR.compare(l.getKey(), r.getKey()))
-                                 .map(e -> {
-                                     joinConditionPartnerTableAliases.addAll(e.getKey().joinConditionPartnerTableAliases);
-                                     return e.getValue();
-                                 })
-                                 .collect(Collectors.joining());
-        return join;
+        Map<RdbmsJoin, String> joinMap = subJoins.stream().collect(Collectors.toMap(j -> j, j -> j.toSql(subSelectJoinPrefix, coercer, sqlParameters, newPrefixes, false)));
+        return subJoins.stream()
+                    .sorted(new RdbmsJoinComparator(subJoins))
+                    .map(j -> {
+                        joinConditionPartnerTableAliases.addAll(j.getJoinConditionPartnerTableAliases());
+                        return joinMap.get(j);
+                    })
+                    .collect(Collectors.joining());
     }
 
 }
